+++ crates/chronik-server/src/kafka_handler.rs
@@ -205,12 +205,12 @@ impl KafkaProtocolHandler {
                 self.protocol_handler.handle_metadata(header, &mut buf).await
             }
             ApiKey::Produce => {
-                eprintln!("!!! PRODUCE HANDLER CALLED !!! version={}, correlation_id={}", header.api_version, header.correlation_id);
+                tracing::debug!("Produce handler called: version={}, correlation_id={}", header.api_version, header.correlation_id);
 
                 // Parse the produce request
                 let request = self.protocol_handler.parse_produce_request(&header, &mut buf)?;
 
-                eprintln!("!!! PRODUCE REQUEST PARSED !!! topics_count={}", request.topics.len());
+                tracing::debug!("Produce request parsed: topics_count={}", request.topics.len());
 
                 // Auto-create topics if they don't exist
                 let topic_names: Vec<String> = request.topics.iter()
@@ -218,7 +218,6 @@ impl KafkaProtocolHandler {
                     .collect();
 
                 if !topic_names.is_empty() {
-                    eprintln!("!!! AUTO-CREATING TOPICS: {:?}", topic_names);
                     tracing::info!("Produce request for topics: {:?}", topic_names);
 
                     // Check which topics don't exist and create them